requirements.txt
Django==1.10.3
django-braces==1.10.0
django-cors-middleware==1.3.1
django-oauth-plus==2.2.9
django-oauth-toolkit==0.10.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.5.3
djangorestframework-oauth==1.1.0
httplib2==0.9.2
oauth2==1.9.0.post1
oauthlib==1.0.3
psycopg2==2.6.2
shortuuid==0.4.3
six==1.10.0
